Remove lines that Rcpp 1.0.7 choked on
Turns out that it was cloning a vector for some early checking.  This was no longer needed, so I could just toss it.  Seems to be working now.

Much in debt to valgrind on my cluster which helped me find the line that was throwing seg faults with the new version of Rcpp.